requirements.txt
pandas
pyyaml
ollama
matplotlib
